+++ src/jk_logging/debugging/__init__.py
@@ -1,11 +1,9 @@
 ﻿
 
 
-__version__ = "0.2021.11.25"
+__version__ = "0.2021.12.8"
 
 
 
 from .DebugTimeStampFormatter import DebugTimeStampFormatter
-
-