+++ src/jk_logging/fmt/__init__.py
@@ -1,7 +1,7 @@
 ﻿
 
 
-__version__ = "0.2021.11.25"
+__version__ = "0.2021.12.8"
 
 
 
@@ -14,6 +14,4 @@ from .ColoredLogMessageFormatter import ColoredLogMessageFormatter, COLOR_LOG_ME
 from .HTMLLogMessageFormatter import HTMLLogMessageFormatter, HTML_LOG_MESSAGE_FORMATTER
 from .JSONLogMessageFormatter import JSONLogMessageFormatter, JSON_LOG_MESSAGE_FORMATTER
 from .LogMessageFormatter import LogMessageFormatter, DEFAULT_LOG_MESSAGE_FORMATTER
-
-